+++ lib/views/settings_screen/components/settings_section.dart
@@ -1,11 +1,11 @@
 import 'package:fluent_ui/fluent_ui.dart';
 
-class FluentSettingsBlock extends StatelessWidget {
+class SettingsSection extends StatelessWidget {
 
   final String title;
   final List<Widget> settings;
 
-  const FluentSettingsBlock({
+  const SettingsSection({
     super.key,
     required this.title,
     required this.settings,
@@ -13,13 +13,12 @@ class FluentSettingsBlock extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    FluentThemeData theme = FluentTheme.of(context);
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 16.0),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          Text(title, style: theme.typography.subtitle),
+          Text(title, style: FluentTheme.of(context).typography.subtitle),
           const SizedBox(height: 8),
           ...settings,
         ],