+++ lib/views/settings_screen/pages/settings_screen_view.face_recognition.dart
@@ -4,14 +4,14 @@ Widget _getFaceRecognitionSettings(SettingsScreenViewModel viewModel, SettingsSc
   return SettingsPage(
     title: "Face recognition",
     blocks: [
-      BooleanInputCard(
+      SettingsToggleTile(
         icon: LucideIcons.scanFace,
         title: "Enable face recognition",
         subtitle: "If enabled, the application allows face recognition features to be used.",
         value: () => viewModel.faceRecognitionEnabled,
         onChanged: controller.onFaceRecognitionEnableChanged,
       ),
-      TextInputCard(
+      SettingsTextEditTile(
         icon: LucideIcons.server,
         title: "Server address",
         subtitle: "The address of the server to connect to for face recognition.",