+++ lib/views/settings_screen/pages/settings_screen_view.output.dart
@@ -4,16 +4,16 @@ Widget _getOutputSettings(SettingsScreenViewModel viewModel, SettingsScreenContr
   return SettingsPage(
     title: "Output",
     blocks: [
-      FluentSettingsBlock(
+      SettingsSection(
         title: "Local",
         settings: [
           Text("Collages are saved to the Output folder of your project.")
         ],
       ),
-      FluentSettingsBlock(
+      SettingsSection(
         title: "Share using internet",
         settings: [
-          TextInputCard(
+          SettingsTextEditTile(
             icon: LucideIcons.globe,
             title: "Firefox Send URL",
             subtitle: "Firefox Send Server URL",
@@ -22,10 +22,10 @@ Widget _getOutputSettings(SettingsScreenViewModel viewModel, SettingsScreenContr
           ),
         ],
       ),
-      FluentSettingsBlock(
+      SettingsSection(
         title: "Image settings",
         settings: [
-          ComboBoxCard(
+          SettingsComboBoxTile(
             icon: LucideIcons.fileImage,
             title: "Image file type",
             subtitle: "What kind of file to generate",
@@ -33,14 +33,14 @@ Widget _getOutputSettings(SettingsScreenViewModel viewModel, SettingsScreenContr
             value: () => viewModel.exportFormat,
             onChanged: controller.onExportFormatChanged,
           ),
-          NumberInputCard(
+          SettingsNumberEditTile(
             icon: LucideIcons.fileSliders,
             title: "JPG quality",
             subtitle: 'Export quality (higher is bigger files)',
             value: () => viewModel.jpgQuality,
             onFinishedEditing: controller.onJpgQualityChanged,
           ),
-          NumberInputCard(
+          SettingsNumberEditTile(
             icon: LucideIcons.proportions,
             title: "Output resolution multiplier",
             subtitle: 'Controls image resolution',